requirements.txt
discord.py==1.7.3
requests==2.25.1
pymongo==3.11.4
aiohttp==3.7.4.post0
Cybernator==0.6.3
asyncio==3.4.3